Move scan methods to Traverse file
